subcompositor: move+rename subsurface_parent_commit()
