+++ include/types/wlr_subcompositor.h
@@ -1,0 +1,8 @@
+#ifndef TYPES_WLR_SUBCOMPOSITOR_H
+#define TYPES_WLR_SUBCOMPOSITOR_H
+
+#include <wlr/types/wlr_subcompositor.h>
+
+void subsurface_handle_parent_commit(struct wlr_subsurface *subsurface);
+
+#endif
